// src/server/include/network/events/types.ts
export enum AltEventType {
    Player_Connect = "playerConnect",
    Player_Disconnect = "playerDisconnect",
    Player_Death = "playerDeath",
    Player_Damage = "playerDamage"
}

export enum ExtendedAltEvent {
    Vehicle_Create = "vehicleCreate"
}

export enum PlayerEvent {
    Connect = "Framework::Player::Connect",
    Disconnect = "Framework::Player::Disconnect",
    Death = "Framework::Player::Death",
    Damage = "Framework::Player::Damage"
}

export enum VehicleEvent {
    Create = "Framework::Vehicle::Create"
}